优化代码

+++ smart-http-test/src/test/java/org/smartboot/http/test/client/HttpPostTest.java
@@ -74,7 +74,7 @@ public class HttpPostTest {
                     bodyBuffer = attachment.get(bodyKey);
                 }
                 if (bodyBuffer == null) {
-                    bodyBuffer = ByteBuffer.allocate(request.getContentLength());
+                    bodyBuffer = ByteBuffer.allocate((int) request.getContentLength());
                     if (attachment == null) {
                         attachment = new Attachment();
                         request.setAttachment(attachment);
